requirements.txt
# MCP File Manager Server Dependencies

# MCP SDK
mcp>=1.0.0

# File operations
aiofiles>=23.2.1
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Data handling
pyyaml>=6.0.1
chardet>=5.2.0

# Security
cryptography>=41.0.0

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.1

# SSE Server (HTTP/HTTPS için)
aiohttp>=3.9.0
aiohttp-sse>=2.1.0

# Image processing (GPT Vision support)
Pillow>=10.0.0

